src/scrape_article/requirements.txt
numpy==1.21.2
pandas==1.3.2
PyVirtualDisplay==2.2
selenium==3.8.0
pdfminer.six==20201018
PyPDF2==1.26.0
